ToEpochInMS -> use Double

+++ src/PGMS.CQSLight/PGMS.IntegratedTests/DataProvider.EFCore/Services/EntityRepositoryFixtures/FindFirstAsyncFixture.cs
@@ -44,7 +44,7 @@ public class FindFirstAsyncFixture
             entityRepository.Insert(new PersonReporting { AggregateRootId = QADataProvider.DwightSchruteId, Name = "Dwight" });
         }
 
-        unixTimeStampMs = DateTime.Now.ToEpochInMilliseconds();
+        unixTimeStampMs = (long)DateTime.Now.ToEpochInMilliseconds();
         entityRepository.Insert(new LogEntryReporting { PersonId = QADataProvider.AngelaMartinId, UnixTimeStampMs = unixTimeStampMs, LocationId = QADataProvider.LocationId });
         entityRepository.Insert(new LogEntryReporting { PersonId = QADataProvider.CreedBrattonId, UnixTimeStampMs = unixTimeStampMs, LocationId = QADataProvider.LocationId });
         entityRepository.Insert(new LogEntryReporting { PersonId = QADataProvider.DwightSchruteId, UnixTimeStampMs = unixTimeStampMs, LocationId = QADataProvider.LocationId });
